load settings, save settings
Settings now save through the `DeforumArgs` method `save()` and the paramater format is saved in yaml instead of json so that it is easier to read for users, the file is saved in .txt for the same reason.

Settings are loaded by importing the .txt file of the saved settings. They  get turned into class variables with SimpleNamespace, and then these variables are updated with the main  `DeforumArgs` class.