init_image batching
Batches init images by entering a directory path into the `init_image` variable instead of an image path (both methods can be used in the same paramater field).
It iterates over the `init_array` list that contains the image paths. If `use_init` is False, the list will countain an empty string so that it runs the prompts loop once without any initial images (so as if it would usually).